refactor(api): split the api/main.py god-file into core + routers
api/main.py had grown to 1040 lines mixing app setup, session state, a
model catalog, helpers, and every endpoint. Split into focused modules
with no behaviour change (all 32 routes register identically; suite green).

* api/core.py: the app, CORS, logging setup, the in-memory session store
  and get_state(), the model catalog and parse/label/availability helpers.
* api/routers/: one APIRouter per pipeline stage, sessions, config,
  analysis, repair, verification, results, each importing shared state
  from core. None exceeds ~216 lines.
* api/main.py: now a 116-line assembler that includes the routers and
  mounts the favicon + frontend static files, preserving the
  favicon-before-static ordering. Re-exports sessions and the helper
  names for back-compat so existing imports keep working.

Fixes uncovered by the split:
* sample-data path now resolves from the qallm package root, not a
  relative hop that broke when the endpoint moved one level deeper.
* test_api_jobs patches _run_verification_work at its new home in the
  verification router (the function genuinely moved).

main.py: 1040 -> 116 lines. Suite: 376 passed.
