+++ src/qallm/api/routers/__init__.py
@@ -1,0 +1,1 @@
+"""API routers, split from the former monolithic main.py."""
