+++ src/qallm/api/routers/config.py
@@ -1,0 +1,178 @@
+"""API router: config endpoints.
+
+Split from the former monolithic api/main.py. Shares the app's session
+store and helpers via qallm.api.core. Behaviour is unchanged; only the
+file boundary moved.
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+import os
+import shutil
+import tempfile
+import uuid
+import zipfile
+from datetime import datetime
+from pathlib import Path
+from typing import List, Optional
+
+from fastapi import APIRouter, File, Form, HTTPException, UploadFile
+from fastapi.responses import FileResponse
+
+from qallm.api.core import (
+    MODEL_CATALOG,
+    check_model_available,
+    get_state,
+    model_label,
+    parse_model_id,
+    sessions,
+)
+from qallm.config import settings
+from qallm.jobs import JobConflictError, JobStatus, get_store
+from qallm.orchestrator import QALLMOrchestrator, LLM_PROVIDERS
+
+logger = logging.getLogger(__name__)
+
+router = APIRouter()
+
+
+@router.get("/api/llm/providers")
+async def get_providers():
+    return {"configured": list(LLM_PROVIDERS.keys())}
+
+
+@router.get("/api/verification/strategies")
+async def get_strategies():
+    return {"configured": ["feedback", "oneshot", "hypothesis"]}
+
+
+@router.get("/api/verification/models")
+async def get_models():
+    """Returns model catalog with availability status for the frontend dropdown."""
+    result = []
+    for entry in MODEL_CATALOG:
+        available = check_model_available(entry)
+        result.append({
+            "id": entry["id"],
+            "label": entry["label"],
+            "provider": entry["provider"],
+            "available": available,
+        })
+    return {"models": result}
+
+
+@router.get("/api/sample-data")
+async def get_sample_data():
+    """Return bundled sample code so users can try the pipeline instantly.
+
+    Each sample is a small module of research-style functions that pass
+    static analysis but contain runtime logic bugs, demonstrating the
+    verification gap. The frontend submits the chosen sample's content
+    through the normal upload flow, so it runs the same path as a real
+    upload.
+    """
+    import os
+
+    import qallm
+    sample_dir = os.path.join(os.path.dirname(qallm.__file__), "sample_data")
+    samples = []
+    descriptions = {
+        "buggy_research_code.py": (
+            "Five functions that look clean to Bandit and Radon but each "
+            "hides a runtime logic bug (off-by-one, aliasing, wrong "
+            "normalisation, a guard that misses zero, a reversed sort). "
+            "Ideal for showing what execution-based verification catches "
+            "that static analysis misses."
+        ),
+    }
+    try:
+        for name in sorted(os.listdir(sample_dir)):
+            if not name.endswith(".py") or name == "__init__.py":
+                continue
+            path = os.path.join(sample_dir, name)
+            with open(path, encoding="utf-8") as fh:
+                content = fh.read()
+            samples.append({
+                "name": name,
+                "description": descriptions.get(name, ""),
+                "content": content,
+                "lines": content.count("\n") + 1,
+            })
+    except OSError:
+        pass
+    return {"samples": samples}
+
+
+@router.get("/api/quality-profiles")
+async def get_quality_profiles():
+    """List selectable quality models (profiles) with explanations.
+
+    Powers the pipeline-start selector. Each profile maps to a quality
+    framework; QALLM evaluates the named indicators per dimension. Only
+    profiles actually implemented are marked available; planned framework
+    profiles are listed so the roadmap is visible in the UI.
+    """
+    from qallm.profiles import _BUILT_IN
+
+    FRAMEWORK = {
+        "implementation_default": "EVERSE Research Software Quality",
+    }
+    FRIENDLY = {
+        "implementation_default": "EVERSE (implementation stage)",
+    }
+    EXPLAIN = {
+        "implementation_default": (
+            "The EVERSE Research Software Quality framework, the v1 "
+            "demonstrator. Evaluates Maintainability, Security, Reliability, "
+            "Reproducibility, and FAIRness using static tools plus QALLM's "
+            "execution-based verification. Thresholds are tuned for the "
+            "implementation stage of the software lifecycle."
+        ),
+    }
+
+    profiles = []
+    for pid, profile in _BUILT_IN.items():
+        profiles.append({
+            "id": pid,
+            "name": FRIENDLY.get(pid, pid),
+            "framework": FRAMEWORK.get(pid, "Custom"),
+            "description": EXPLAIN.get(pid, profile.description),
+            "available": True,
+            "dimensions": [
+                {"name": d.dimension.value,
+                 "indicators": [i.name for i in d.indicators]}
+                for d in profile.dimensions
+            ],
+        })
+
+    # Roadmap entries: declared so the selector shows where QALLM is going,
+    # but not selectable until implemented.
+    roadmap = [
+        {"id": "iso25010", "name": "ISO/IEC 25010", "framework": "ISO/IEC 25010",
+         "description": "Software product quality model. Planned as an "
+                        "alternative profile; not yet implemented.",
+         "available": False, "dimensions": []},
+        {"id": "fair4rs", "name": "FAIR4RS", "framework": "FAIR for Research Software",
+         "description": "FAIR principles specialised for research software. "
+                        "Planned; FAIRness indicators already exist and would "
+                        "anchor this profile.",
+         "available": False, "dimensions": []},
+    ]
+
+    return {"profiles": profiles + roadmap, "default": "implementation_default"}
+
+
+@router.get("/api/analysis/tools")
+async def get_analysis_tools():
+    return {"tools": ["bandit", "radon", "ruff", "trufflehog"]}
+
+
+@router.get("/api/session/{session_id}/config")
+async def get_session_config(session_id: str):
+    """Returns the locked-in session configuration (model, strategy, oracle)."""
+    state = get_state(session_id)
+    return {"config": state.get("config", {})}
+
+
